requirements.txt
fastapi
uvicorn
langchain
tavily-python
yfinance
openai
yfinance